+++ packages/effect/src/unstable/sql/SqlModel.ts
@@ -21,11 +21,13 @@ import * as SqlSchema from "./SqlSchema.ts"
  */
 export const makeRepository = <
   S extends Model.Any,
-  Id extends (keyof S["Type"]) & (keyof S["update"]["Type"]) & (keyof S["fields"])
+  Id extends (keyof S["Type"]) & (keyof S["update"]["Type"]) & (keyof S["fields"]),
+  SoftDelete extends keyof S["fields"] = never
 >(Model: S, options: {
   readonly tableName: string
   readonly spanPrefix: string
   readonly idColumn: Id
+  readonly softDeleteColumn?: SoftDelete | undefined
 }): Effect.Effect<
   {
     readonly insert: (
@@ -66,6 +68,12 @@ export const makeRepository = <
     const sql = yield* SqlClient
     const idSchema = Model.fields[options.idColumn] as Schema.Top
     const idColumn = options.idColumn as string
+    const softDeleteColumn = options.softDeleteColumn as string | undefined
+    const withSoftDeleteFilter = (where: any) =>
+      softDeleteColumn === undefined ? where : sql.and([where, sql`${sql(softDeleteColumn)} is null`])
+    const setSoftDeleted = softDeleteColumn === undefined
+      ? undefined
+      : sql`${sql(softDeleteColumn)} = CURRENT_TIMESTAMP`
 
     const insertSchema = SqlSchema.findOne({
       Request: Model.insert,
@@ -74,9 +82,10 @@ export const makeRepository = <
         sql.onDialectOrElse({
           mysql: () =>
             sql`insert into ${sql(options.tableName)} ${sql.insert(request as any)};
-select * from ${sql(options.tableName)} where ${sql(idColumn)} = LAST_INSERT_ID();`.unprepared.pipe(
-              Effect.map(([, results]) => results as any)
-            ),
+select * from ${sql(options.tableName)} where ${withSoftDeleteFilter(sql`${sql(idColumn)} = LAST_INSERT_ID()`)};`
+              .unprepared.pipe(
+                Effect.map(([, results]) => results as any)
+              ),
           orElse: () => sql`insert into ${sql(options.tableName)} ${sql.insert(request as any).returning("*")}`
         })
     })
@@ -111,15 +120,16 @@ select * from ${sql(options.tableName)} where ${sql(idColumn)} = LAST_INSERT_ID(
       execute: (request: any) =>
         sql.onDialectOrElse({
           mysql: () =>
-            sql`update ${sql(options.tableName)} set ${sql.update(request, [idColumn])} where ${sql(idColumn)} = ${
-              request[idColumn]
+            sql`update ${sql(options.tableName)} set ${sql.update(request, [idColumn])} where ${
+              withSoftDeleteFilter(sql`${sql(idColumn)} = ${request[idColumn]}`)
             };
-select * from ${sql(options.tableName)} where ${sql(idColumn)} = ${request[idColumn]};`.unprepared.pipe(
-              Effect.map(([, results]) => results as any)
-            ),
+select * from ${sql(options.tableName)} where ${withSoftDeleteFilter(sql`${sql(idColumn)} = ${request[idColumn]}`)};`
+              .unprepared.pipe(
+                Effect.map(([, results]) => results as any)
+              ),
           orElse: () =>
-            sql`update ${sql(options.tableName)} set ${sql.update(request, [idColumn])} where ${sql(idColumn)} = ${
-              request[idColumn]
+            sql`update ${sql(options.tableName)} set ${sql.update(request, [idColumn])} where ${
+              withSoftDeleteFilter(sql`${sql(idColumn)} = ${request[idColumn]}`)
             } returning *`
         })
     })
@@ -142,8 +152,8 @@ select * from ${sql(options.tableName)} where ${sql(idColumn)} = ${request[idCol
     const updateVoidSchema = SqlSchema.void({
       Request: Model.update,
       execute: (request: any) =>
-        sql`update ${sql(options.tableName)} set ${sql.update(request, [idColumn])} where ${sql(idColumn)} = ${
-          request[idColumn]
+        sql`update ${sql(options.tableName)} set ${sql.update(request, [idColumn])} where ${
+          withSoftDeleteFilter(sql`${sql(idColumn)} = ${request[idColumn]}`)
         }`
     })
     const updateVoid = (
@@ -160,7 +170,8 @@ select * from ${sql(options.tableName)} where ${sql(idColumn)} = ${request[idCol
     const findByIdSchema = SqlSchema.findOne({
       Request: idSchema,
       Result: Model,
-      execute: (id: any) => sql`select * from ${sql(options.tableName)} where ${sql(idColumn)} = ${id}`
+      execute: (id: any) =>
+        sql`select * from ${sql(options.tableName)} where ${withSoftDeleteFilter(sql`${sql(idColumn)} = ${id}`)}`
     })
     const findById = (
       id: S["fields"][Id]["Type"]
@@ -177,7 +188,12 @@ select * from ${sql(options.tableName)} where ${sql(idColumn)} = ${request[idCol
 
     const deleteSchema = SqlSchema.void({
       Request: idSchema,
-      execute: (id: any) => sql`delete from ${sql(options.tableName)} where ${sql(idColumn)} = ${id}`
+      execute: (id: any) =>
+        softDeleteColumn === undefined
+          ? sql`delete from ${sql(options.tableName)} where ${sql(idColumn)} = ${id}`
+          : sql`update ${sql(options.tableName)} set ${setSoftDeleted} where ${
+            withSoftDeleteFilter(sql`${sql(idColumn)} = ${id}`)
+          }`
     })
     const delete_ = (
       id: S["fields"][Id]["Type"]
@@ -201,13 +217,15 @@ select * from ${sql(options.tableName)} where ${sql(idColumn)} = ${request[idCol
  */
 export const makeResolvers = <
   S extends Model.Any,
-  Id extends (keyof S["Type"]) & (keyof S["update"]["Type"]) & (keyof S["fields"])
+  Id extends (keyof S["Type"]) & (keyof S["update"]["Type"]) & (keyof S["fields"]),
+  SoftDelete extends keyof S["fields"] = never
 >(
   Model: S,
   options: {
     readonly tableName: string
     readonly spanPrefix: string
     readonly idColumn: Id
+    readonly softDeleteColumn?: SoftDelete | undefined
   }
 ): Effect.Effect<
   {
@@ -246,6 +264,12 @@ export const makeResolvers = <
     const sql = yield* SqlClient
     const idSchema = Model.fields[options.idColumn] as Schema.Top
     const idColumn = options.idColumn as string
+    const softDeleteColumn = options.softDeleteColumn as string | undefined
+    const withSoftDeleteFilter = (where: any) =>
+      softDeleteColumn === undefined ? where : sql.and([where, sql`${sql(softDeleteColumn)} is null`])
+    const setSoftDeleted = softDeleteColumn === undefined
+      ? undefined
+      : sql`${sql(softDeleteColumn)} = CURRENT_TIMESTAMP`
 
     const insert: RequestResolver.RequestResolver<
       SqlResolver.SqlRequest<
@@ -262,9 +286,10 @@ export const makeResolvers = <
           mysql: () =>
             Effect.forEach(request, (request: any) =>
               sql`insert into ${sql(options.tableName)} ${sql.insert(request)};
-select * from ${sql(options.tableName)} where ${sql(idColumn)} = LAST_INSERT_ID();`.unprepared.pipe(
-                Effect.map(([, results]) => results[0] as any)
-              ), { concurrency: 10 }),
+select * from ${sql(options.tableName)} where ${withSoftDeleteFilter(sql`${sql(idColumn)} = LAST_INSERT_ID()`)};`
+                .unprepared.pipe(
+                  Effect.map(([, results]) => results[0] as any)
+                ), { concurrency: 10 }),
           orElse: () => sql`insert into ${sql(options.tableName)} ${sql.insert(request).returning("*")}`
         })
     }).pipe(
@@ -293,7 +318,8 @@ select * from ${sql(options.tableName)} where ${sql(idColumn)} = LAST_INSERT_ID(
       ResultId(request: any) {
         return request[idColumn]
       },
-      execute: (ids: any) => sql`select * from ${sql(options.tableName)} where ${sql.in(idColumn, ids)}`
+      execute: (ids: any) =>
+        sql`select * from ${sql(options.tableName)} where ${withSoftDeleteFilter(sql.in(idColumn, ids))}`
     }).pipe(
       RequestResolver.withSpan(`${options.spanPrefix}.findByIdResolver`)
     )
@@ -307,7 +333,12 @@ select * from ${sql(options.tableName)} where ${sql(idColumn)} = LAST_INSERT_ID(
       >
     > = SqlResolver.void({
       Request: idSchema,
-      execute: (ids: any) => sql`delete from ${sql(options.tableName)} where ${sql.in(idColumn, ids)}`
+      execute: (ids: any) =>
+        softDeleteColumn === undefined
+          ? sql`delete from ${sql(options.tableName)} where ${sql.in(idColumn, ids)}`
+          : sql`update ${sql(options.tableName)} set ${setSoftDeleted} where ${
+            withSoftDeleteFilter(sql.in(idColumn, ids))
+          }`
     }).pipe(
       RequestResolver.withSpan(`${options.spanPrefix}.deleteResolver`)
     )