+++ packages/effect/test/schema/Schema.test.ts
@@ -34,8 +34,6 @@ import { produce } from "immer"
 import { deepStrictEqual, fail, ok, strictEqual } from "node:assert"
 import { assertFalse, assertInclude, assertTrue, throws } from "../utils/assert.ts"
 
-const isDeno = "Deno" in globalThis
-
 const verifyGeneration = true
 
 const equals = TestSchema.Asserts.ast.fields.equals
@@ -1977,6 +1975,9 @@ Expected a value with a size of at most 2, got Map([["a",1],["b",NaN],["c",3]])`
       }
 
       const zoned = DateTime.makeZonedUnsafe("2021-01-01T00:00:00.000Z", { timeZone: "Europe/London" })
+
+      const decoding = asserts.decoding()
+      await decoding.fail("invalid", `Invalid Zoned DateTime string: invalid`)
 
       const encoding = asserts.encoding()
       await encoding.succeed(zoned, DateTime.formatIsoZoned(zoned))
@@ -4190,6 +4191,7 @@ Expected a value with a size of at most 2, got Map([["a",1],["b",NaN],["c",3]])`
 
     const decoding = asserts.decoding()
     await decoding.succeed("2021-01-01T00:00:00.000Z", DateTime.makeUnsafe("2021-01-01T00:00:00.000Z"))
+    await decoding.fail("invalid", `Invalid UTC DateTime string: invalid`)
     await decoding.fail(null, `Expected string, got null`)
 
     const encoding = asserts.encoding()
@@ -4469,6 +4471,25 @@ Expected a value with a size of at most 2, got Map([["a",1],["b",NaN],["c",3]])`
     }
   })
 
+  it("DurationFromString", async () => {
+    const schema = Schema.DurationFromString
+    const asserts = new TestSchema.Asserts(schema)
+
+    const decoding = asserts.decoding()
+    await decoding.succeed("1000 millis", Duration.millis(1000))
+    await decoding.succeed("1 second", Duration.seconds(1))
+    await decoding.succeed("Infinity", Duration.infinity)
+    await decoding.succeed("-Infinity", Duration.negativeInfinity)
+    await decoding.fail("value", "Invalid Duration string: value")
+
+    const encoding = asserts.encoding()
+    await encoding.succeed(Duration.zero, "0 millis")
+    await encoding.succeed(Duration.seconds(5), "5000 millis")
+    await encoding.succeed(Duration.nanos(5000n), "5000 nanos")
+    await encoding.succeed(Duration.infinity, "Infinity")
+    await encoding.succeed(Duration.negativeInfinity, "-Infinity")
+  })
+
   it("DurationFromNanos", async () => {
     const schema = Schema.DurationFromNanos
     const asserts = new TestSchema.Asserts(schema)
@@ -4718,7 +4739,7 @@ Expected a value with a size of at most 2, got Map([["a",1],["b",NaN],["c",3]])`
     await decoding.succeed("https://effect.website", new URL("https://effect.website"))
     await decoding.fail(
       "123",
-      isDeno ? `TypeError: Invalid URL: '123'` : `TypeError: Invalid URL`
+      `Invalid URL string: 123`
     )
 
     const encoding = asserts.encoding()